refactor(diffusion): consolidate MiniMax H3 VAE modules
