+++ python/sglang/multimodal_gen/runtime/models/vaes/minimax_h3_audio_vae/__init__.py
@@ -1,5 +1,5 @@
 # SPDX-License-Identifier: Apache-2.0
 
-from .dac_audio_vae import DacAudioVAE
+from .audio_vae import DacAudioVAE
 
 __all__ = ["DacAudioVAE"]
