--- python/sglang/multimodal_gen/runtime/models/vaes/minimax_h3_audio_vae/dac_utils.py
@@ -1,12 +1,0 @@
-# SPDX-License-Identifier: MIT
-# Adapted from https://github.com/jik876/hifi-gan under the MIT license.
-
-
-def init_weights(m, mean=0.0, std=0.01):
-    classname = m.__class__.__name__
-    if classname.find("Conv") != -1:
-        m.weight.data.normal_(mean, std)
-
-
-def get_padding(kernel_size, dilation=1):
-    return int((kernel_size * dilation - dilation) / 2)
