+++ python/sglang/multimodal_gen/runtime/models/vaes/minimax_h3_video_vae/klvae.py
@@ -6,6 +6,7 @@ from typing import List, Union
 
 import numpy as np
 import torch
+import torch.distributed as dist
 import torch.nn as nn
 from diffusers.configuration_utils import ConfigMixin, register_to_config
 from diffusers.loaders.single_file_model import FromOriginalModelMixin
@@ -15,13 +16,17 @@ from PIL import Image
 
 from sglang.multimodal_gen.runtime.distributed import (
     get_decode_parallel_group_coordinator,
+    get_decode_parallel_rank,
+    get_decode_parallel_world_size,
+    model_parallel_is_initialized,
 )
 
-from .normalize import get_denormalize_transform, get_normalize_transform
-from .parallel import get_tile_parallel_state
+from .processor import (
+    VAEProcessor,
+    get_denormalize_transform,
+    get_normalize_transform,
+)
 from .vae_cnn import EncoderFCN3D
-from .vae_module import ClsTokenAggregator, DiagonalGaussianDistribution
-from .vae_processor import VAEProcessor
 from .vae_vit import ViT3DDecoder
 
 logger = logging.get_logger(__name__)  # pylint: disable=invalid-name
@@ -57,6 +62,61 @@ def _resolve_temporal_stream_cat():
         .lower()
     )
     return raw not in ("0", "false", "no", "off", "disable", "disabled")
+
+
+def get_tile_parallel_state():
+    if not dist.is_initialized() or not model_parallel_is_initialized():
+        return 0, 1
+    return get_decode_parallel_rank(), get_decode_parallel_world_size()
+
+
+class DiagonalGaussianDistribution(object):
+    def __init__(self, parameters, upcast_fp32=True):
+        if upcast_fp32:
+            parameters = parameters.to(dtype=torch.float32)
+
+        self.parameters = parameters
+        self.mean, self.logvar = torch.chunk(parameters, 2, dim=1)
+        self.logvar = torch.clamp(self.logvar, -30.0, 20.0)
+        self.std = self.logvar.mul(0.5).exp_()
+
+    @torch.compiler.disable
+    def sample(self, generator=None):
+        noise = torch.randn(self.mean.shape, generator=generator)
+        noise = noise.to(device=self.parameters.device)
+        return noise.mul_(self.std).add_(self.mean)
+
+
+class ClsTokenAggregator:
+    def __init__(self, vae_model):
+        self.vae = vae_model
+        self.cls_tokens = []
+
+    def __enter__(self):
+        return self
+
+    def __exit__(self, exc_type, exc_val, exc_tb):
+        if self.cls_tokens and hasattr(self.vae.encoder, "loss_info"):
+            self.vae.encoder.loss_info["cls_token"] = torch.stack(
+                self.cls_tokens, dim=0
+            ).mean(dim=0)
+        return False
+
+    def collect(self):
+        if (
+            hasattr(self.vae.encoder, "loss_info")
+            and "cls_token" in self.vae.encoder.loss_info
+        ):
+            self.cls_tokens.append(self.vae.encoder.loss_info["cls_token"].clone())
+
+    def collect_stacked(self, num_tiles, sample_batch_size):
+        if (
+            hasattr(self.vae.encoder, "loss_info")
+            and "cls_token" in self.vae.encoder.loss_info
+        ):
+            cls_token = self.vae.encoder.loss_info["cls_token"]
+            cls_token = cls_token.unflatten(0, (num_tiles, sample_batch_size))
+            self.cls_tokens.extend(token.clone() for token in cls_token)
 
 
 class AutoencoderKL(ModelMixin, ConfigMixin, FromOriginalModelMixin):
