+++ python/sglang/multimodal_gen/runtime/models/vaes/minimax_h3_video_vae/processor.py
@@ -1,13 +1,54 @@
 # SPDX-License-Identifier: Apache-2.0
 # Tensor pre/post-processing for the MiniMax H3 visual VAE.
 import math
+from typing import Tuple
 
 import numpy as np
 import torch
 from diffusers.utils import logging
 from einops import rearrange
+from torchvision.transforms import Normalize
 
-from .normalize import get_denormalize_transform, get_normalize_transform
+NORM_CONFIGS = {
+    "imagenet": {
+        "mean": (0.485, 0.456, 0.406),
+        "std": (0.229, 0.224, 0.225),
+    },
+    "simple": {
+        "mean": (0.5, 0.5, 0.5),
+        "std": (0.5, 0.5, 0.5),
+    },
+    "raw": {
+        "mean": (0.0, 0.0, 0.0),
+        "std": (1.0, 1.0, 1.0),
+    },
+}
+
+
+def get_norm_constants(
+    norm_type: str = "imagenet",
+) -> Tuple[Tuple[float, ...], Tuple[float, ...]]:
+    if norm_type not in NORM_CONFIGS:
+        raise ValueError(
+            f"Unknown norm_type: {norm_type}. Must be one of {list(NORM_CONFIGS.keys())}"
+        )
+    config = NORM_CONFIGS[norm_type]
+    return config["mean"], config["std"]
+
+
+def get_normalize_transform(
+    norm_type: str = "imagenet", *, inplace: bool = False
+) -> Normalize:
+    mean, std = get_norm_constants(norm_type)
+    return Normalize(mean, std, inplace=inplace)
+
+
+def get_denormalize_transform(norm_type: str = "imagenet") -> Normalize:
+    mean, std = get_norm_constants(norm_type)
+    inv_mean = tuple(-m / s for m, s in zip(mean, std))
+    inv_std = tuple(1.0 / s for s in std)
+    return Normalize(inv_mean, inv_std)
+
 
 logger = logging.get_logger(__name__)  # pylint: disable=invalid-name
